+++ src/index.ts
@@ -9,6 +9,6 @@ export * from "./ScSet";
 export * from "./ScTemplate";
 export * from "./ScTemplateResult";
 export * from "./ScType";
-export * from "./ScHelpers";
+export * from "./ScHelper";
 export * from "./constants";
 export * from "./errors";